requirements.txt
requests
Flask
nba_api
pytz
python-dotenv
psycopg2-binary
supabase
playwright
beautifulsoup4
redis
flask-compress
orjson
